feat: creating last services tests
