+++ services/questionService.ts
@@ -1,31 +1,17 @@
-import {
-    getFirestore,
-    collection,
-    getDocs,
-    QuerySnapshot,
-    DocumentData,
-} from 'firebase/firestore'
-import { firebaseApp } from '../firebaseConfig'
+import { collection, doc, getDoc, getDocs } from 'firebase/firestore'
+import { db } from '@/firebaseConfig'
 import Question from '@/dtos/Question'
 
-const db = getFirestore(firebaseApp)
-
-/**
- * Fetches all documents from the 'Question' collection in Firestore.
- * @returns {Promise<Question[]>} A promise that resolves to an array of quiz questions.
- * @throws Will throw an error if fetching fails.
- */
-export const fetchAllQuestions = async (): Promise<Question[]> => {
+export const fetchAllQuestionsService = async (): Promise<{
+    questions: Question[]
+} | null> => {
     try {
         const questionsCollectionRef = collection(db, 'Question')
-        const querySnapshot: QuerySnapshot<DocumentData> = await getDocs(
-            questionsCollectionRef
-        )
+        const questionsSnapshot = await getDocs(questionsCollectionRef)
 
-        const questions: Question[] = []
-        querySnapshot.forEach((doc) => {
+        const questions: Question[] = questionsSnapshot.docs.map((doc) => {
             const data = doc.data()
-            questions.push({
+            return {
                 id: doc.id,
                 answers: data.answers || [],
                 description: data.description || 'No description',
@@ -33,10 +19,35 @@ export const fetchAllQuestions = async (): Promise<Question[]> => {
                 // image: data.image,
                 rightAnswer: data.rightAnswer,
                 title: data.title || 'Untitled Question',
-            } as Question)
+            } as Question
         })
-        return questions
+        return { questions }
     } catch (error) {
         throw new Error('Failed to fetch quiz questions.')
+    }
+}
+
+export const fetchQuestionById = async(questionId: string): Promise<Question> => {
+    try {
+        const questionDocRef = doc(db, 'Question', questionId)
+        const questionSnapshot = await getDoc(questionDocRef)
+
+        if (questionSnapshot.exists()) {
+            const data = questionSnapshot.data()
+            return {
+                id: questionSnapshot.id,
+                answers: data.answers || [],
+                description: data.description || 'No description',
+                explanation: data.explanation || 'No explanation',
+                // image: data.image,
+                rightAnswer: data.rightAnswer,
+                title: data.title || 'Untitled Question',
+            } as Question
+        } else {
+            throw new Error(`Question with ID ${questionId} does not exist.`)
+        }
+    } catch (error) {
+        console.error('Error fetching question:', error)
+        throw new Error('Failed to fetch the question.')
     }
 }
