+++ stores/lessonStore.ts
@@ -2,7 +2,7 @@ import { create } from 'zustand'
 import Lesson from '@/dtos/Lesson'
 import { getLessonByIdService, markLessonAsConcluded } from '@/services/lessonService'
 import Question from '@/dtos/Question'
-import { fetchQuestion } from '@/services/questionsService'
+import { fetchQuestion } from '@/services/questionService'
 
 interface LessonState {
     lesson: Lesson | null
